+++ src/client/app/character/note/note.service.spec.ts
@@ -12,7 +12,8 @@ describe('NoteService', () => {
     TestBed.configureTestingModule({
       imports: [HttpClientModule, HttpClientTestingModule],
       providers: [NoteService]
-    }));
+    })
+  );
 
   it('should be created', () => {
     const service: NoteService = TestBed.get(NoteService);