+++ src/client/app/models/races/zora.spec.ts
@@ -1,35 +1,47 @@
+import { Attributes } from '#Enums/attributes.enum';
 import { Zora } from '#Races/zora';
+
+const zora = 'Zora';
 
 test('should create a Zora', () => {
   const myZora = new Zora();
   expect(myZora).toBeTruthy();
-  expect(myZora.race).toBe('Zora');
+  expect(myZora.race).toBe(zora);
   expect(myZora.attributes[1].value).toBe(10);
 });
-test('should create a River Zora', () => {
-  const myZora = new Zora('River');
+test('should create a River Zora', () =>
+  subRaceTest(
+    'River',
+    { name: 'Constitution', value: 6 },
+    { name: 'Intelligence', value: 10 }
+  ));
+test('should create a Ocean Zora', () =>
+  subRaceTest(
+    'Ocean',
+    { name: 'Strength', value: 10 },
+    { name: 'Wisdom', value: 6 }
+  ));
+test('should create a Swamp Zora', () =>
+  subRaceTest(
+    'Swamp',
+    { name: 'Strength', value: 6 },
+    { name: 'Constitution', value: 10 }
+  ));
+
+function subRaceTest(
+  subRace: string,
+  attribute1: { name: string; value: number },
+  attribute2: { name: string; value: number }
+): any {
+  const myZora = new Zora(subRace);
   expect(myZora).toBeTruthy();
-  expect(myZora.race).toBe('Zora');
-  expect(myZora.subRace).toBe('River');
+  expect(myZora.race).toBe(zora);
+  expect(myZora.subRace).toBe(subRace);
   expect(myZora.attributes[1].value).toBe(10);
-  expect(myZora.attributes[2].value).toBe(6);
-  expect(myZora.attributes[3].value).toBe(10);
-});
-test('should create a Ocean Zora', () => {
-  const myZora = new Zora('Ocean');
-  expect(myZora).toBeTruthy();
-  expect(myZora.race).toBe('Zora');
-  expect(myZora.subRace).toBe('Ocean');
-  expect(myZora.attributes[1].value).toBe(10);
-  expect(myZora.attributes[0].value).toBe(10);
-  expect(myZora.attributes[4].value).toBe(6);
-});
-test('should create a Swamp Zora', () => {
-  const myZora = new Zora('Swamp');
-  expect(myZora).toBeTruthy();
-  expect(myZora.race).toBe('Zora');
-  expect(myZora.subRace).toBe('Swamp');
-  expect(myZora.attributes[1].value).toBe(10);
-  expect(myZora.attributes[0].value).toBe(6);
-  expect(myZora.attributes[2].value).toBe(10);
-});
+  expect(myZora.attributes[Attributes[attribute1.name]].value).toBe(
+    attribute1.value
+  );
+  expect(myZora.attributes[Attributes[attribute2.name]].value).toBe(
+    attribute2.value
+  );
+}