+++ src/server/app/utils/utils.ts
@@ -1,34 +1,23 @@
-const base62Charset = Array.from(
-  '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'
-);
+export class Utils {
+  static base62Array: string[] = Array.from(
+    '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'
+  );
 
-/**
- * A function that takes in a length and makes a random string that many character long from 0-9a-Z
- * @param length how many characters long to make the random string
- * @returns {string} id - The generated string that is returned
- */
-export function makeId(length): string {
-  let id = '';
-  for (let i = 0; i < length; i++) {
-    id += base62Charset[Math.round(Math.random() * 100) % 62];
+  static checkNull(value: string | number, type: string): string | number {
+    if (type === 'string') {
+      return value ? value : '';
+    } else if (type === 'number') {
+      return value === 0 || value ? value : null;
+    } else {
+      return undefined;
+    }
   }
-  return id;
-}
 
-/**
- * A function to check if a value is null or undefined and return a blank string or the value 'null'
- * @param {string | number }value the value that needs to be checked if null or undefined
- * @returns {number | string} a blank string or the value null
- */
-export function checkNull(
-  value: string | number,
-  type: string
-): string | number {
-  if (type === 'string') {
-    return value ? value : '';
-  } else if (type === 'number') {
-    return value ? value : null;
-  } else {
-    return '';
+  static makeId(length): string {
+    let id = '';
+    for (let i = 0; i < length; i++) {
+      id += Utils.base62Array[Math.round(Math.random() * 100) % 62];
+    }
+    return id;
   }
 }